update constants folder
